+++ .gitignore
@@ -4,4 +4,19 @@ npm-debug.log*
 node_modules
 .npm
 .DS_Store
-dist
+
+# TypeScript
+*.tsbuildinfo
+
+# Testing
+coverage/
+.nyc_output/
+
+# IDE
+.vscode/
+.idea/
+*.swp
+*.swo
+
+# OS
+Thumbs.db
