+++ package.json
@@ -7,7 +7,9 @@
   "scripts": {
     "build": "tsc",
     "dev": "tsc --watch",
-    "test": "npm run build && jasmine",
+    "test": "npm run build && vitest run",
+    "test:watch": "vitest",
+    "test:coverage": "vitest run --coverage",
     "lint": "standard",
     "lint-fix": "standard --fix",
     "prepublishOnly": "npm run build"
@@ -23,12 +25,10 @@
   },
   "homepage": "https://github.com/masterT/bandcamp-scraper",
   "devDependencies": {
-    "@types/jasmine": "^4.3.0",
     "@types/node": "^20.0.0",
-    "jasmine": "^2.8.0",
     "standard": "^16.0.3",
-    "ts-node": "^10.9.0",
-    "typescript": "^5.0.0"
+    "typescript": "^5.0.0",
+    "vitest": "^1.6.1"
   },
   "dependencies": {
     "ajv": "^4.11.8",
